extract immatriculation history to own component
